# Transition shape when object's capacity changes