… ext/-test-/rational/depend
… ext/pty/depend
… object.c
… shape.c
… shape.h
… test/-ext-/string/test_cstr.rb
… test/objspace/test_objspace.rb
… test/ruby/test_gc_compact.rb
… test/ruby/test_io_buffer.rb
… test/ruby/test_shapes.rb
… variable.c
… vm.c
… vm_core.h
… vm_insnhelper.c
… yjit/src/codegen.rs
… yjit/src/cruby_bindings.inc.rs