This PR introduces shape transitions on capacity changes, and includes 3 commits:

* Omit an IO Buffer test on 32 bit machines which demonstrates a bug ([#19084](https://bugs.ruby-lang.org/issues/19084)) when strings are extended (currently only testing embedded strings). This commit should be reverted when the bug is fixed
* Transition shapes when an object's capacity changes. This includes when objects are allocated out of bigger size pools. Notably, after this commit, any two objects with the same shape will have the same capacity
* Remove the `numiv` field from objects. Since an object's shape will now give us the capacity, we can remove this field from objects, and instead use it to embed one more instance variable