@maximecb 2022-11-09
Happy to see that this is now working 👍 

Out of curiosity: do you try to guess the correct size pool of objects when allocating them based on the number of ivars in the class? It seems like this would be a useful optimization to minimize shape transitions. My main concern is that multiple shape transitions could be bad for YJIT since we might generate code that is no longer useful. This is something that could be optimized in later PRs and doesn't necessarily have to be optimal right away, but I'm curious to know nevertheless.