@jemmaissroff 2022-11-09
> Out of curiosity: do you try to guess the correct size pool of objects when allocating them based on the number of ivars in the class? It seems like this would be a useful optimization to minimize shape transitions. My main concern is that multiple shape transitions could be bad for YJIT since we might generate code that is no longer useful. This is something that could be optimized in later PRs and doesn't necessarily have to be optimal right away, but I'm curious to know nevertheless.

@maximecb yes, this was happening before this PR! [We use the `max_iv_count` on a class to pick which size pool to allocate an object out of.](https://github.com/ruby/ruby/blob/master/gc.c#L2919-L2926) And the `max_iv_count` [is incremented on a class](https://github.com/ruby/ruby/blob/master/shape.c#L153-L155) if we get a new shape for an object with that class whose number of ivars is bigger than `max_iv_count.` The only slight "gotcha" is that this will miss subclasses, so [we also set the `max_iv_count` while garbage collecting](https://github.com/ruby/ruby/blob/master/gc.c#L7280-L7284) to do a best effort at this case. All of this means that _most_ of the time, if we know an object is likely to be in a bigger size pool, it will only do the initial transition from the root shape to its appropriate size pool shape, and wouldn't do multiple shape transitions. Also worth noting that the root shape is also the shape for the default size pool, so new object allocations where we know nothing about the class won't do any initial transitions on allocation.