@k0kubun 2022-11-15
I noticed that this slightly slowed down YJIT as well:

```
before: ruby 3.2.0dev (2022-11-10T15:11:34Z master 9986697b62) +YJIT [arm64-darwin22]
after: ruby 3.2.0dev (2022-11-10T15:11:34Z master c726c48a3d) +YJIT [arm64-darwin22]

-----  -----------  ----------  ----------  ----------  ------------  -------------
bench  before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
nbody  45.6         0.6         47.4        0.4         0.96          0.96
-----  -----------  ----------  ----------  ----------  ------------  -------------
```

<details>

```
$ ./run_benchmarks.rb nbody -e "before::/opt/rubies/yjit-release-before-$arch/bin/ruby --yjit" -e "after::/opt/rubies/yjit-release-$arch/bin/ruby --yjit"
Running benchmark "nbody" (1/1)
/opt/rubies/yjit-release-before-arm64/bin/ruby --yjit -I ./harness benchmarks/nbody/benchmark.rb
ruby 3.2.0dev (2022-11-10T15:11:34Z master 9986697b62) +YJIT [arm64-darwin22]
itr #1: 47ms
itr #2: 45ms
itr #3: 45ms
itr #4: 45ms
itr #5: 45ms
itr #6: 45ms
itr #7: 45ms
itr #8: 45ms
itr #9: 45ms
itr #10: 45ms
itr #11: 45ms
itr #12: 45ms
itr #13: 45ms
itr #14: 45ms
itr #15: 45ms
itr #16: 45ms
itr #17: 45ms
itr #18: 45ms
itr #19: 45ms
itr #20: 45ms
itr #21: 45ms
itr #22: 45ms
itr #23: 45ms
itr #24: 45ms
itr #25: 45ms
itr #26: 45ms
itr #27: 45ms
itr #28: 45ms
itr #29: 45ms
itr #30: 45ms
itr #31: 45ms
itr #32: 45ms
itr #33: 45ms
itr #34: 45ms
itr #35: 45ms
itr #36: 45ms
itr #37: 45ms
itr #38: 45ms
itr #39: 45ms
itr #40: 45ms
itr #41: 47ms
itr #42: 46ms
itr #43: 46ms
itr #44: 45ms
itr #45: 45ms
itr #46: 45ms
itr #47: 45ms
itr #48: 45ms
itr #49: 45ms
itr #50: 45ms
itr #51: 45ms
itr #52: 45ms
itr #53: 45ms
itr #54: 45ms
itr #55: 45ms
itr #56: 45ms
itr #57: 45ms
itr #58: 45ms
itr #59: 45ms
itr #60: 45ms
itr #61: 45ms
itr #62: 45ms
itr #63: 45ms
itr #64: 45ms
itr #65: 45ms
itr #66: 45ms
itr #67: 45ms
itr #68: 45ms
itr #69: 45ms
itr #70: 45ms
itr #71: 45ms
itr #72: 45ms
itr #73: 45ms
itr #74: 45ms
itr #75: 45ms
itr #76: 45ms
itr #77: 45ms
itr #78: 45ms
itr #79: 45ms
itr #80: 45ms
itr #81: 45ms
itr #82: 45ms
itr #83: 47ms
itr #84: 45ms
itr #85: 46ms
itr #86: 45ms
itr #87: 46ms
itr #88: 45ms
itr #89: 45ms
itr #90: 45ms
itr #91: 45ms
itr #92: 45ms
itr #93: 45ms
itr #94: 45ms
itr #95: 45ms
itr #96: 45ms
itr #97: 45ms
itr #98: 46ms
itr #99: 45ms
itr #100: 45ms
itr #101: 45ms
itr #102: 45ms
itr #103: 45ms
itr #104: 45ms
itr #105: 45ms
itr #106: 45ms
itr #107: 45ms
itr #108: 45ms
itr #109: 45ms
itr #110: 45ms
itr #111: 45ms
itr #112: 45ms
itr #113: 45ms
itr #114: 45ms
itr #115: 45ms
itr #116: 45ms
itr #117: 45ms
itr #118: 45ms
itr #119: 45ms
itr #120: 45ms
itr #121: 45ms
itr #122: 45ms
itr #123: 45ms
itr #124: 45ms
itr #125: 45ms
itr #126: 45ms
itr #127: 45ms
itr #128: 45ms
itr #129: 46ms
itr #130: 45ms
itr #131: 45ms
itr #132: 45ms
itr #133: 45ms
itr #134: 45ms
itr #135: 45ms
itr #136: 45ms
itr #137: 45ms
itr #138: 45ms
itr #139: 45ms
itr #140: 45ms
itr #141: 45ms
itr #142: 45ms
itr #143: 45ms
itr #144: 45ms
itr #145: 45ms
itr #146: 45ms
itr #147: 45ms
itr #148: 45ms
itr #149: 45ms
itr #150: 45ms
itr #151: 45ms
itr #152: 45ms
itr #153: 45ms
itr #154: 45ms
itr #155: 45ms
itr #156: 45ms
itr #157: 45ms
itr #158: 45ms
itr #159: 45ms
itr #160: 45ms
itr #161: 45ms
itr #162: 45ms
itr #163: 45ms
itr #164: 45ms
itr #165: 45ms
itr #166: 45ms
itr #167: 45ms
itr #168: 45ms
itr #169: 45ms
itr #170: 45ms
itr #171: 45ms
itr #172: 45ms
itr #173: 47ms
itr #174: 45ms
itr #175: 45ms
itr #176: 45ms
itr #177: 45ms
itr #178: 45ms
itr #179: 45ms
itr #180: 45ms
itr #181: 45ms
itr #182: 45ms
itr #183: 45ms
itr #184: 45ms
itr #185: 45ms
itr #186: 45ms
itr #187: 45ms
itr #188: 45ms
itr #189: 45ms
itr #190: 45ms
itr #191: 45ms
itr #192: 45ms
itr #193: 45ms
itr #194: 45ms
itr #195: 45ms
itr #196: 45ms
itr #197: 45ms
itr #198: 45ms
itr #199: 45ms
itr #200: 45ms
itr #201: 45ms
itr #202: 45ms
itr #203: 45ms
itr #204: 45ms
itr #205: 45ms
itr #206: 45ms
itr #207: 45ms
itr #208: 45ms
itr #209: 45ms
itr #210: 45ms
itr #211: 45ms
itr #212: 45ms
itr #213: 45ms
itr #214: 45ms
itr #215: 45ms
itr #216: 45ms
itr #217: 46ms
itr #218: 45ms
itr #219: 45ms
itr #220: 45ms
Average of last 205, non-warmup iters: 45ms
Running benchmark "nbody" (1/1)
/opt/rubies/yjit-release-arm64/bin/ruby --yjit -I ./harness benchmarks/nbody/benchmark.rb
ruby 3.2.0dev (2022-11-10T15:11:34Z master c726c48a3d) +YJIT [arm64-darwin22]
itr #1: 49ms
itr #2: 47ms
itr #3: 47ms
itr #4: 47ms
itr #5: 47ms
itr #6: 47ms
itr #7: 47ms
itr #8: 47ms
itr #9: 47ms
itr #10: 47ms
itr #11: 47ms
itr #12: 47ms
itr #13: 47ms
itr #14: 47ms
itr #15: 47ms
itr #16: 47ms
itr #17: 47ms
itr #18: 47ms
itr #19: 47ms
itr #20: 47ms
itr #21: 47ms
itr #22: 47ms
itr #23: 47ms
itr #24: 47ms
itr #25: 47ms
itr #26: 47ms
itr #27: 47ms
itr #28: 47ms
itr #29: 47ms
itr #30: 47ms
itr #31: 47ms
itr #32: 47ms
itr #33: 47ms
itr #34: 47ms
itr #35: 47ms
itr #36: 47ms
itr #37: 47ms
itr #38: 47ms
itr #39: 47ms
itr #40: 47ms
itr #41: 47ms
itr #42: 47ms
itr #43: 47ms
itr #44: 47ms
itr #45: 47ms
itr #46: 47ms
itr #47: 47ms
itr #48: 47ms
itr #49: 47ms
itr #50: 47ms
itr #51: 47ms
itr #52: 47ms
itr #53: 47ms
itr #54: 47ms
itr #55: 47ms
itr #56: 47ms
itr #57: 47ms
itr #58: 47ms
itr #59: 47ms
itr #60: 47ms
itr #61: 47ms
itr #62: 47ms
itr #63: 47ms
itr #64: 47ms
itr #65: 47ms
itr #66: 47ms
itr #67: 47ms
itr #68: 47ms
itr #69: 47ms
itr #70: 47ms
itr #71: 47ms
itr #72: 47ms
itr #73: 47ms
itr #74: 47ms
itr #75: 47ms
itr #76: 47ms
itr #77: 47ms
itr #78: 47ms
itr #79: 47ms
itr #80: 47ms
itr #81: 47ms
itr #82: 47ms
itr #83: 47ms
itr #84: 47ms
itr #85: 47ms
itr #86: 47ms
itr #87: 47ms
itr #88: 47ms
itr #89: 47ms
itr #90: 47ms
itr #91: 47ms
itr #92: 47ms
itr #93: 47ms
itr #94: 47ms
itr #95: 47ms
itr #96: 47ms
itr #97: 47ms
itr #98: 47ms
itr #99: 47ms
itr #100: 47ms
itr #101: 47ms
itr #102: 47ms
itr #103: 47ms
itr #104: 47ms
itr #105: 47ms
itr #106: 47ms
itr #107: 47ms
itr #108: 47ms
itr #109: 47ms
itr #110: 47ms
itr #111: 47ms
itr #112: 47ms
itr #113: 47ms
itr #114: 47ms
itr #115: 47ms
itr #116: 48ms
itr #117: 47ms
itr #118: 47ms
itr #119: 47ms
itr #120: 47ms
itr #121: 47ms
itr #122: 47ms
itr #123: 47ms
itr #124: 47ms
itr #125: 47ms
itr #126: 47ms
itr #127: 47ms
itr #128: 47ms
itr #129: 47ms
itr #130: 47ms
itr #131: 47ms
itr #132: 47ms
itr #133: 47ms
itr #134: 47ms
itr #135: 47ms
itr #136: 47ms
itr #137: 47ms
itr #138: 47ms
itr #139: 47ms
itr #140: 47ms
itr #141: 47ms
itr #142: 47ms
itr #143: 47ms
itr #144: 47ms
itr #145: 47ms
itr #146: 47ms
itr #147: 47ms
itr #148: 47ms
itr #149: 47ms
itr #150: 47ms
itr #151: 47ms
itr #152: 47ms
itr #153: 47ms
itr #154: 47ms
itr #155: 47ms
itr #156: 47ms
itr #157: 47ms
itr #158: 48ms
itr #159: 47ms
itr #160: 47ms
itr #161: 47ms
itr #162: 47ms
itr #163: 47ms
itr #164: 47ms
itr #165: 47ms
itr #166: 47ms
itr #167: 47ms
itr #168: 47ms
itr #169: 47ms
itr #170: 47ms
itr #171: 47ms
itr #172: 47ms
itr #173: 47ms
itr #174: 47ms
itr #175: 47ms
itr #176: 47ms
itr #177: 47ms
itr #178: 47ms
itr #179: 47ms
itr #180: 47ms
itr #181: 47ms
itr #182: 47ms
itr #183: 47ms
itr #184: 47ms
itr #185: 47ms
itr #186: 47ms
itr #187: 47ms
itr #188: 47ms
itr #189: 47ms
itr #190: 47ms
itr #191: 47ms
itr #192: 47ms
itr #193: 47ms
itr #194: 47ms
itr #195: 47ms
itr #196: 47ms
itr #197: 47ms
itr #198: 47ms
itr #199: 47ms
itr #200: 48ms
itr #201: 47ms
itr #202: 47ms
itr #203: 47ms
itr #204: 47ms
itr #205: 47ms
itr #206: 47ms
itr #207: 47ms
itr #208: 47ms
itr #209: 47ms
itr #210: 47ms
itr #211: 47ms
Average of last 196, non-warmup iters: 47ms
Total time spent benchmarking: 20s

end_time: 2022-11-15 15:14:55 PST (-0800)
before: ruby 3.2.0dev (2022-11-10T15:11:34Z master 9986697b62) +YJIT [arm64-darwin22]
after: ruby 3.2.0dev (2022-11-10T15:11:34Z master c726c48a3d) +YJIT [arm64-darwin22]

-----  -----------  ----------  ----------  ----------  ------------  -------------
bench  before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
nbody  45.6         0.6         47.4        0.4         0.96          0.96
-----  -----------  ----------  ----------  ----------  ------------  -------------
Legend:
- before/after: ratio of before/after time. Higher is better for after. Above 1 represents a speedup.
- after 1st itr: ratio of before/after time for the first benchmarking iteration.
```

</details>